Some improvement in a better accro mode, I hope

+++ Libraries/AQ_Gyroscope/Gyroscope.h
@@ -26,9 +26,9 @@
 
 #define FINDZERO 49
 
-int gyroADC[3] = {0,0,0};
+long  gyroADC[3] = {0,0,0};
 float gyroRate[3] = {0.0,0.0,0.0};
-int   gyroZero[3] = {0,0,0};
+long  gyroZero[3] = {0,0,0};
 long  gyroSample[3] = {0,0,0};
 float gyroScaleFactor = 0.0;
 float gyroOneMeterSecADCFactor = 0.0;
